move stream extensions
