+++ Orchestrator.Core/Interfaces/IEnvelopeStreamService.cs
@@ -12,28 +12,4 @@ namespace Orchestrator.Core.Interfaces
         IAsyncEnumerable<Envelope> StreamAsync(string topic);
     }
 
-    public static class EnvelopeStreamServiceExtensions
-    {
-        /// <summary>“Unwrap” the envelope, yielding the strongly-typed payload.</summary>
-        public static async IAsyncEnumerable<T> StreamAsync<T>(
-            this IEnvelopeStreamService svc,
-            string topic)
-        {
-            await foreach (var envelope in svc.StreamAsync(topic))
-            {
-                yield return envelope.Payload.Deserialize<T>()!;
-            }
-        }
-
-        /// <summary>Gives you the raw JSON string of the envelope payload.</summary>
-        public static async IAsyncEnumerable<string> StreamRawAsync(
-            this IEnvelopeStreamService svc,
-            string topic)
-        {
-            await foreach (var envelope in svc.StreamAsync(topic))
-            {
-                yield return envelope.Payload.GetRawText();
-            }
-        }
-    }
 }
